requirements.txt
python-telegram-bot>=21.0
requests==2.31.0
httpx==0.27.0
pytz